Split filter test file
